+++ src/ambition/Ambition_unit.cc
@@ -32,6 +32,7 @@
 #define _AMBITION_IMPLEMENTATION
 #include "OFIRM.h"
 #include "OREMOTE.h"
+#include "OTOWN.h"
 #include "OUNIT.h"
 
 #include "Ambition_building.hh"
@@ -168,6 +169,17 @@ void Unit::sendToBuildingRallyPoint(
 ) {
   const auto building
     = Building::findBy7kaaFirmRecordNumber(_7kaaFirm->firm_recno);
+  if (building) {
+    building->sendUnitsToRallyPoint(_7kaaUnitRecordNumbers);
+  }
+}
+
+void Unit::sendToBuildingRallyPoint(
+  std::vector<short> _7kaaUnitRecordNumbers,
+  const Town* _7kaaTown
+) {
+  const auto building
+    = Building::findBy7kaaTownRecordNumber(_7kaaTown->town_recno);
   if (building) {
     building->sendUnitsToRallyPoint(_7kaaUnitRecordNumbers);
   }